+++ src/components/pages/CheckinCalendar/useCheckinCalendar.ts
@@ -1,11 +1,10 @@
-import { useFoursquare } from '@/hooks/useFoursquare'
 import { useRecoil } from '@/hooks/useRecoil'
 import { getStartEndOfMonth } from '@/service/dateFns'
+import { fetchUserCheckins, fetchUser } from '@/service/foursquareApi'
 import { generateImageUrl } from '@/service/utilFns'
 import { useState, useEffect } from 'react'
 
 export const useCheckinCalendar = () => {
-  const { fetchUserCheckins, fetchUser } = useFoursquare()
   const [loading, setLoading] = useState(true)
   const { setUser, setCheckins, setFetchHistory, fetchHistory } = useRecoil()
   const [userProfURL, setUserProfURL] = useState<string>()
