+++ src/components/pages/SignInByFoursquare.tsx
@@ -1,15 +1,14 @@
 import { config } from '@/service/config'
 import React from 'react'
 import { WebView, WebViewNavigation } from 'react-native-webview'
-import { useFoursquare } from '@/hooks/useFoursquare'
 import { NavigationProp, ParamListBase, useNavigation } from '@react-navigation/native'
 import { FOURSQUARE_ACCESS_TOKEN } from '@/constants/StorageKeys'
 import { setUserId, logEvent } from '@/hooks/useAnalytics'
 import storage from '@/service/reactNativeStorage'
 import { parseURLParams } from '@/service/utilFns'
+import { fetchAccessToken, fetchUser } from '@/service/foursquareApi'
 
 const SignInByFoursquare = () => {
-  const { fetchAccessToken, fetchUser } = useFoursquare()
   const { CLIENT_ID, REDIRECT_URI } = config()
   const navigation = useNavigation<NavigationProp<ParamListBase>>()
