+++ src/components/pages/UserProfile.tsx
@@ -2,11 +2,11 @@ import React, { useEffect } from 'react'
 import { View, ActivityIndicator } from 'react-native'
 import { useRecoil } from '@/hooks/useRecoil'
 import { useNavigation, RouteProp } from '@react-navigation/native'
-import { useFoursquare } from '@/hooks/useFoursquare'
 import UserCard from '@/components/card/UserCard'
 import useAsyncFn from 'react-use/lib/useAsyncFn'
 import { RootStackParamList } from '@/types'
 import { other } from '@/styles/styles'
+import { fetchUser } from '@/service/foursquareApi'
 
 type Props = {
   route: RouteProp<RootStackParamList, 'UserProfile'>
@@ -16,7 +16,6 @@ const UserProfile = ({ route }: Props) => {
   const userId = route.params?.userId
   const navigation = useNavigation()
   const { setUser } = useRecoil()
-  const { fetchUser } = useFoursquare()
   const [userTemp, fetchUserTemp] = useAsyncFn(async () => await fetchUser(userId), [])
 
   useEffect(() => {
