+++ src/service/utilFns.ts
@@ -1,4 +1,4 @@
-import { Checkin } from '@/types/Foursquare'
+import { Checkin, FoursquareResponse } from '@/types/Foursquare'
 import { AgendaItemsMap } from 'react-native-calendars'
 import { getDateString } from './dateFns'
 import 'react-native-url-polyfill/auto'
@@ -72,4 +72,17 @@ export const parseURLParams = (rawUrl: string, key: string) => {
   if (params.has(key)) {
     return params.get(key)
   }
+}
+
+type Props = {
+  res: void | Response
+  type: 'checkins' | 'checkin' | 'user'
+}
+
+export const responseExtractor = async <T>({ res, type }: Props): Promise<T> => {
+  const parsedRes = (await (res as Response).json()) as FoursquareResponse
+  if (parsedRes.meta.code !== 200) {
+    console.error({ error: 'failed', message: parsedRes.meta.errorDetail })
+  }
+  return parsedRes.response[type] as unknown as T
 }
